requirements.txt
streamlit
torch
torchvision
pillow
flask
